+++ src/index.ts
@@ -2,4 +2,4 @@ export * as calendar from './lib/api/calendar';
 export * as youtube from './lib/api/youtube';
 export { getItems } from './lib/api/shared';
 export { createProfile, getProfiles } from './lib/profiles';
-export { login } from './lib/auth';
+export { login, getAuth } from './lib/auth';