+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.0](../../tags/v3.0.0) - 2023-03-13
+### Changed
+- Revised auth instructions
+- Scopes can be set per end-project
+
 ## [2.1.3](../../tags/v2.1.3) - 2023-03-13
 ### Changed
 - Fixed exported types